requirements/development.txt
# Development dependencies
-r base.txt

# Code formatting
black==24.3.0
isort==5.13.2

# Linting
flake8==7.0.0
pylint==3.1.0
pylint-django==2.5.5

# Type checking
mypy==1.9.0
django-stubs==4.2.7

# Development tools
ipython==8.22.2
django-extensions==3.2.3

# Debugging
# django-debug-toolbar==4.3.0

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==24.3.0
